feat(recall): convert synapt to PEP 420 implicit namespace package (Step 3 of synapt.premium rename chain)
Drops src/synapt/__init__.py so synapt becomes a PEP 420 namespace package. This enables synapt.premium (and future synapt.runner, synapt.<extension>) to coexist under the synapt.* namespace without name collision on the root __init__.

Load-bearing for premium#716 deployability per Sentinel's blocker #1 (clean install + import synapt.premium fails because current public synapt owns synapt/__init__.py — coexistence broken).

Changes:
- Delete src/synapt/__init__.py (was the only thing blocking PEP 420)
- Move __version__ to src/synapt/recall/__init__.py (lives with recall as the OSS surface that owns it)
- Update 5 import sites: from synapt import __version__ → from synapt.recall import __version__
- Update src/synapt/server.py _find_watch_paths to iterate synapt.__path__ for namespace package
- Update scripts/bump-version.sh to target recall/__init__.py

Premium boundary: OSS recall — package structure change to enable namespace coexistence; no premium logic in OSS.

Coordinated with: premium#716 (Apollo author-merges after this lands + publish + premium fixes blockers 2/3/4).
