--- src/synapt/__init__.py
@@ -1,3 +1,0 @@
-"""Synapt: persistent conversational memory for AI coding assistants."""
-
-__version__ = "0.15.1"
